requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
coverage==5.5
Deprecated==1.2.13
deprecation==2.1.0
Django==4.0
django-environ==0.8.1
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-recaptcha2==1.4.1
django-selectel-storage==1.0.2
djangorestframework==3.13.1
eyed3==0.9.6
filetype==1.0.9
gunicorn==20.1.0
idna==3.3
packaging==21.3
Pillow==8.4.0
psycopg2==2.9.2
pyparsing==3.0.6
python-slugify==5.0.2
pytz==2021.3
redis==4.1.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
wrapt==1.13.3
